requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
matplotlib==3.7.1
Pillow==10.0.0
tkinter-tooltip==1.7.0
scipy==1.11.1
scikit-learn==1.3.0
joblib==1.3.2 